dont bother moving/copying when directory exists

+++ build.bash
@@ -121,8 +121,12 @@ export device=athene
 releasetype=experimental
 
 # setup
-cp -r ~/android/fireclaw722/android_device_motorola_athene/ ~/android/lineage/oreo-mr1/device/motorola/athene
-cp -r ~/android/fireclaw722/proprietary_vendor_motorola/ ~/android/lineage/oreo-mr1/vendor/motorola
+if [ ! -d "~/android/lineage/oreo-mr1/device/motorola/athene" ]; then
+        cp -r ~/android/fireclaw722/android_device_motorola_athene/ ~/android/lineage/oreo-mr1/device/motorola/athene
+fi
+if [ ! -d "~/android/lineage/oreo-mr1/vendor/motorola" ]; then
+        cp -r ~/android/fireclaw722/proprietary_vendor_motorola/ ~/android/lineage/oreo-mr1/vendor/motorola
+fi
 
 # run build
 cd ~/android/lineage/oreo-mr1
@@ -133,8 +137,12 @@ saveFiles
 
 # cleanup
 cleanMka
-rm -rf ~/android/lineage/oreo-mr1/device/motorola/athene
-rm -rf ~/android/lineage/oreo-mr1/vendor/motorola
+if [ -d "~/android/lineage/oreo-mr1/device/motorola/athene" ]; then
+        rm -rf ~/android/lineage/oreo-mr1/device/motorola/athene
+fi
+if [ -d "~/android/lineage/oreo-mr1/vendor/motorola" ]; then
+        rm -rf ~/android/lineage/oreo-mr1/vendor/motorola
+fi
 
 #ONEPLUS3
 export device=oneplus3
@@ -178,8 +186,12 @@ export device=addison
 releasetype=release
 
 # setup
-cp -r ~/android/fireclaw722/device_motorola_addison/ ~/android/lineage/oreo-mr1/device/motorola/addison
-cp -r ~/android/fireclaw722/vendor_motorola/ ~/android/lineage/oreo-mr1/vendor/motorola
+if [ ! -d "~/android/lineage/oreo-mr1/device/motorola/addison" ]; then
+        cp -r ~/android/fireclaw722/device_motorola_addison/ ~/android/lineage/oreo-mr1/device/motorola/addison
+fi
+if [ ! -d "~/android/lineage/oreo-mr1/vendor/motorola" ]; then
+        cp -r ~/android/fireclaw722/vendor_motorola/ ~/android/lineage/oreo-mr1/vendor/motorola
+fi
 
 # run build
 cd ~/android/lineage/oreo-mr1
@@ -190,5 +202,9 @@ saveFiles
 
 # cleanup
 cleanMka
-rm -rf ~/android/lineage/oreo-mr1/device/motorola/addison
-rm -rf ~/android/lineage/oreo-mr1/vendor/motorola
+if [ -d "~/android/lineage/oreo-mr1/device/motorola/addison" ]; then
+        rm -rf ~/android/lineage/oreo-mr1/device/motorola/addison
+fi
+if [ -d "~/android/lineage/oreo-mr1/vendor/motorola" ]; then
+        rm -rf ~/android/lineage/oreo-mr1/vendor/motorola
+fi
